+++ CHANGELOG.md
@@ -1,1 +1,32 @@
-# Changelog\n\n## [2.0.0] - 2025-02-16\n\n### Added\n- \n\n### Changed\n- \n\n### Fixed\n- \n
+# Changelog
+
+## [1.3.0] - 2025-02-20
+
+### Added
+
+- Dev server to preview templates
+
+### Changed
+
+- Updated invoice template
+- Test suite for templates
+
+## [1.2.0] - 2025-02-20
+
+### Added
+
+- Added Bulgarian translations
+
+### Changed
+
+- Updated invoice template
+
+## [1.0.0] - 2025-02-16
+
+### Added
+
+- \n
+
+### Changed
+
+- \n